Extract scan state manager to a separate class
